+++ django_iban/forms.py
@@ -2,25 +2,54 @@ from django import forms
 from .validators import IBANValidator, swift_bic_validator, IBAN_COUNTRY_CODE_LENGTH
 
 
-iban_min_length = min(IBAN_COUNTRY_CODE_LENGTH.values())
+IBAN_MIN_LENGTH = min(IBAN_COUNTRY_CODE_LENGTH.values())
 
 
 class IBANFormField(forms.CharField):
     """
     An IBAN consists of up to 34 alphanumeric characters.
 
+    To limit validation to specific countries, set the 'include_countries' argument with a tuple or list of ISO 3166-1
+    alpha-2 codes. For example, `include_countries=('NL', 'BE, 'LU')`.
+
+    A list of countries that use IBANs as part of SEPA is included for convenience. To use this feature, set
+    `include_countries=IBAN_SEPA_COUNTRIES` as an argument to the field.
+
+    Example:
+
+    .. code-block:: python
+
+        from django import forms
+        from localflavor.generic.forms import IBANFormField
+        from localflavor.generic.sepa_countries import IBAN_SEPA_COUNTRIES
+
+        class MyForm(forms.Form):
+            iban = IBANFormField(include_countries=IBAN_SEPA_COUNTRIES)
+
+    In addition to validating official IBANs, this field can optionally validate unofficial IBANs that have been
+    catalogued by Nordea by setting the `use_nordea_extensions` argument to True.
+
     https://en.wikipedia.org/wiki/International_Bank_Account_Number
     """
-    default_validators = [IBANValidator()]
-
-    def __init__(self, *args, **kwargs):
-        kwargs.setdefault('min_length', iban_min_length)
+    def __init__(self, use_nordea_extensions=False, include_countries=None, *args, **kwargs):
+        kwargs.setdefault('min_length', IBAN_MIN_LENGTH)
         kwargs.setdefault('max_length', 34)
+        self.default_validators = [IBANValidator(use_nordea_extensions, include_countries)]
         super(IBANFormField, self).__init__(*args, **kwargs)
 
     def to_python(self, value):
         value = super(IBANFormField, self).to_python(value)
-        return value.replace(' ', '')
+        if value is not None:
+            return value.upper().replace(' ', '').replace('-', '')
+        return value
+
+    def prepare_value(self, value):
+        """ The display format for IBAN has a space every 4 characters. """
+        if value is None:
+            return value
+        grouping = 4
+        value = value.upper().replace(' ', '').replace('-', '')
+        return ' '.join(value[i:i + grouping] for i in range(0, len(value), grouping))
 
 
 class SWIFTBICFormField(forms.CharField):